finished all stylings

+++ main.py
@@ -19,95 +19,105 @@ class MainPage:
         # window.configure(bg='')
 
         # set first frame
-        self.frame0 = Frame(window)
-        self.frame0.pack()
+        self.frame0 = Frame(window, bg='yellow')
+        self.frame0.pack(fill="both", expand=True)
 
-        openfile_button = Button(self.frame0, text="Open existing file", cursor="hand2", bg='yellow2', borderwidth='2', relief='groove', font=(
+        openfile_button = Button(self.frame0, text="Open existing file", cursor="hand2", bg='green', fg='white', borderwidth='2', relief='groove', font=(
             'Arial 13'), command=self.openFile)
-        openfile_button.pack()
+        openfile_button.pack(padx=20, pady=0, ipadx=10, ipady=8, anchor='center', side='bottom', expand=True)
+        #anchor='ne', side='bottom'
 
-        createfile_button = Button(self.frame0, text="create new file", cursor="hand2", bg='yellow2', borderwidth='2', relief='groove', font=(
+        createfile_button = Button(self.frame0, text="create new file", cursor="hand2", bg='green', fg='white', borderwidth='2', relief='groove', font=(
             'Arial 13'), command=self.createFile)
-        createfile_button.pack()
+        createfile_button.pack(padx=20, ipadx=20, ipady=8, pady=0, anchor='center', side='top', expand=True)
 
         # set create new file frame with subject title
-        self.frame1 = Frame(window)
+        self.frame1 = Frame(window, bg='yellow')
 
-        subject = Label(self.frame1, text="Enter the subject / course")
-        subject.pack()
+        subject = Label(self.frame1, text="COURSE TITLE", anchor='w', padx=20, pady=25, bg='yellow', fg='black', font=('Arial Bold', 13), justify=LEFT)
+        subject.pack(anchor='w')
 
         # storing the course title in a string
         self.courseInput = StringVar()
 
         # text varible should be stored in the empty string variable created (self.courseInput)
         subjectEntry = Entry(self.frame1, width=30, font=(
-            'Arial 16'), textvariable=self.courseInput)
-        subjectEntry.pack()
+            'Arial 15'), justify=LEFT, highlightthickness=2, relief='groove', textvariable=self.courseInput)
+        subjectEntry.focus()
+        subjectEntry.config(highlightcolor='green4', highlightbackground='green4')
+        subjectEntry.pack(ipadx=10, ipady=10, anchor='w', padx=20)
 
         # added event listener to the button for hiding the initial frame and showing frame2
-        submit_button = Button(self.frame1, text="Submit", cursor="hand2", font=(
+        submit_button = Button(self.frame1, text="Submit", cursor="hand2", bg='green', fg='white', borderwidth='2', relief='groove', font=(
             'Arial 13'), command=self.showFields)
-        submit_button.pack()
+        submit_button.pack(padx=20, ipadx=20, ipady=8, pady=20, anchor='w', side='top')
 
         # for going back to frame0
-        back_button = Button(self.frame1, text="Back", cursor="hand2", font=(
+        back_button = Button(self.frame1, text="Back", cursor="hand2", bg='green', fg='white', borderwidth='2', relief='groove', font=(
             'Arial 13'), command=self.backToFrame0Btn)
-        back_button.pack()
+        back_button.pack(padx=30, pady=30, ipadx=20, ipady=8, anchor='ne', side='bottom')
 
         # third frame with title
-        self.frame2 = Frame(window)
+        self.frame2 = Frame(window, bg='yellow')
+
         # initial title to be changed to the inputted value
         self.courseInputValue = Label(
-            self.frame2, text='Subject Title', font=('Arial 16'))
-        self.courseInputValue.pack()
+            self.frame2, text='Subject Title', font=('Arial 16'), padx=20, pady=10, bg='yellow', fg='black', justify=LEFT)
+        self.courseInputValue.pack(anchor='w')
 
-        matricNo = Label(self.frame2, text="Matric No")
+        matricNo = Label(self.frame2, text="Matric No", bg='yellow', padx=20, pady=10, font=('Arial 13'))
         matricNo.pack()
 
         self.matricNo = StringVar()
         self.MatricNo = Entry(self.frame2, width=30, font=(
-            'Arial 16'), textvariable=self.matricNo)
-        self.MatricNo.pack()
-
-        FirstName = Label(self.frame2, text="First Name")
+            'Arial 15'), highlightthickness=2, relief='flat', textvariable=self.matricNo)
+        self.MatricNo.focus()
+        self.MatricNo.config(highlightcolor='green4', highlightbackground='green4')
+        self.MatricNo.pack(ipadx=10, ipady=3, padx=20)
+       
+        FirstName = Label(self.frame2, text="First Name", bg='yellow', padx=20, pady=10, font=('Arial 13'))
         FirstName.pack()
 
         self.firstname = StringVar()
         self.FirstName = Entry(self.frame2, width=30, font=(
-            'Arial 16'), textvariable=self.firstname)
-        self.FirstName.pack()
+            'Arial 15'), highlightthickness=2, relief='flat', textvariable=self.firstname)
+        self.FirstName.config(highlightcolor='green4', highlightbackground='green4')
+        self.FirstName.pack(ipadx=10, ipady=3, padx=20)
 
-        LastName = Label(self.frame2, text="Last Name")
+        LastName = Label(self.frame2, text="Last Name", bg='yellow', padx=20, pady=10, font=('Arial 13'))
         LastName.pack()
 
         self.lastname = StringVar()
         self.LastName = Entry(self.frame2, width=30, font=(
-            'Arial 16'), textvariable=self.lastname)
-        self.LastName.pack()
+            'Arial 15'), highlightthickness=2, relief='flat', textvariable=self.lastname)
+        self.LastName.config(highlightcolor='green4', highlightbackground='green4')
+        self.LastName.pack(ipadx=10, ipady=3, padx=20)
 
-        Score = Label(self.frame2, text="Score")
+        Score = Label(self.frame2, text="Score", bg='yellow', padx=20, pady=10, font=('Arial 13'))
         Score.pack()
 
         self.score = IntVar()
         self.Score = Entry(self.frame2, width=30, font=(
-            'Arial 16'), textvariable=self.score)
-        self.Score.pack()
+            'Arial 15'), highlightthickness=2, relief='flat', textvariable=self.score)
+        self.Score.config(highlightcolor='green4', highlightbackground='green4')
+        self.Score.pack(ipadx=10, ipady=3, padx=20)
 
         # added event listener to the button to save user info
-        submit_button = Button(self.frame2, text="Save Entry", cursor="hand2", font=(
+        submit_button = Button(self.frame2, text="Save Entry", cursor="hand2", bg='green', fg='white', borderwidth='2', relief='groove', font=(
             'Arial 13'), command=self.saveUserInfos)
-        submit_button.pack()
+        submit_button.pack(padx=20, ipadx=20, ipady=8, pady=20, anchor='w', side='left')
 
         # button to clear entry
-        clear_button = Button(self.frame2, text="Clear Entry", cursor="hand2", font=(
+        clear_button = Button(self.frame2, text="Clear Entry", cursor="hand2", bg='green', fg='white', borderwidth='2', relief='groove', font=(
             'Arial 13'), command=self.clearEntry)
-        clear_button.pack()
+        clear_button.pack(padx=20, ipadx=20, ipady=8, pady=20, anchor='e', side='right')
 
         # menu bar for run analysis
-        self.menubar = Menu(window)
+        self.menubar = Menu(self.frame0, background='red', foreground='white')
         window.config(menu=self.menubar)
 
         self.analysisMenu = Menu(self.menubar, tearoff=0)
+        self.analysisMenu.config(bg = "GREEN",fg='white',activebackground='yellow',activeforeground='black',relief=FLAT)
 
         # I commented this line of code to hide the run analysis button, the code is later in the open file function and the show field function
 
@@ -124,16 +134,16 @@ class MainPage:
             label='Students that failed', command=self.failedScore)
 
         # creating a forth frame for displaying eda results
-        self.frame3 = Frame(window)
-        self.result = Label(self.frame3, text="")
+        self.frame3 = Frame(window, bg='yellow')
+        self.result = Label(self.frame3, text="", font=('Arial 16'), padx=20, pady=10, bg='yellow', fg='black')
         self.result.pack()
-        self.namesLabel = Label(self.frame3, text='')
+        self.namesLabel = Label(self.frame3, text='', font=('Arial 16'), padx=20, pady=4, bg='yellow', fg='black')
         self.namesLabel.pack()
 
         # for going back to frame2
-        back_button = Button(self.frame3, text="Back", cursor="hand2", font=(
+        back_button = Button(self.frame3, text="Back", cursor="hand2", bg='green', fg='white', borderwidth='2', relief='groove', font=(
             'Arial 13'), command=self.backToFrame2Btn)
-        back_button.pack()
+        back_button.pack(padx=30, pady=10, ipadx=20, ipady=8, anchor='ne', side='bottom')
 
         window.mainloop()
 
@@ -142,7 +152,7 @@ class MainPage:
         self.filename = filedialog.askopenfilename(
             initialdir='/', title='Open file', filetypes=(("Json file", "*.json"), ("xml files", "*.xml")))
         if self.filename != '':
-            self.frame2.pack()
+            self.frame2.pack(fill="both", expand=True)
             self.frame0.pack_forget()
 
             # setting subject value from the saved file
@@ -163,23 +173,23 @@ class MainPage:
             initialdir='/', title='Save file', defaultextension=".json")
 
         if self.filename != '':
-            self.frame1.pack()
+            self.frame1.pack(fill="both", expand=True)
             self.frame0.pack_forget()
 
             self.command = 'create file'
 
     # back to frame0 function
     def backToFrame0Btn(self):
-        self.frame0.pack()
+        self.frame0.pack(fill="both", expand=True)
         self.frame1.pack_forget()
 
     def showFields(self):
-        courseTitleInput = self.courseInput.get()
+        courseTitleInput = self.courseInput.get().upper()
 
         if courseTitleInput != "":
 
             # to make the third frame visible
-            self.frame2.pack()
+            self.frame2.pack(fill="both", expand=True)
 
             # to hide the second frame
             self.frame1.pack_forget()
@@ -190,48 +200,52 @@ class MainPage:
 
             # setting the subject value from the subject value input field
             if self.command == 'create file':
-                self.courseInputValue['text'] = self.courseInput.get()
+                self.courseInputValue['text'] = self.courseInput.get().upper()
         else:
-            tkinter.messagebox.showinfo("Input", "Subject field is empty!")
+            tkinter.messagebox.showerror("Error", "Course title field is empty!")
 
     def saveUserInfos(self):
 
-        # create a new row everytime new informations are added
-        # new row will be added to the multi dimensional list, the matrix[] list
-        current_row = [self.courseInput.get(), self.matricNo.get(), self.firstname.get(),
+        if self.matricNo.get() != '' and self.score.get != '':
+            # create a new row everytime new informations are added
+            # new row will be added to the multi dimensional list, the matrix[] list
+            current_row = [self.courseInput.get(), self.matricNo.get(), self.firstname.get(),
                        self.lastname.get(), self.score.get()]
 
-        # appending the new row to the main list
-        self.matrix.append(current_row)
+            # appending the new row to the main list
+            self.matrix.append(current_row)
 
-        if self.command == 'create file':
-            with open(self.filename, 'w') as outfile:
-                json.dump(self.matrix, outfile)
-        elif self.command == 'open file':
-            with open(self.filename, 'r') as openfile:
-                json_object = json.load(openfile)
-                json_object.append(current_row)
+            if self.command == 'create file':
+                with open(self.filename, 'w') as outfile:
+                    json.dump(self.matrix, outfile)
+            elif self.command == 'open file':
+                with open(self.filename, 'r') as openfile:
+                    json_object = json.load(openfile)
+                    json_object.append(current_row)
 
-                # print(json_object)
+                    # print(json_object)
 
-            # updating the newly added informations
-            with open(self.filename, 'w') as outfile:
-                json.dump(json_object, outfile)
+                # updating the newly added informations
+                with open(self.filename, 'w') as outfile:
+                    json.dump(json_object, outfile)
 
-        # clearing the input fields after submission for new entry
-        self.MatricNo.delete(0, END)
-        self.MatricNo.insert(0, '')
-        self.FirstName.delete(0, END)
-        self.FirstName.insert(0, '')
-        self.LastName.delete(0, END)
-        self.LastName.insert(0, '')
-        self.Score.delete(0, END)
-        self.Score.insert(0, '')
+            # clearing the input fields after submission for new entry
+            self.MatricNo.delete(0, END)
+            self.MatricNo.insert(0, '')
+            self.FirstName.delete(0, END)
+            self.FirstName.insert(0, '')
+            self.LastName.delete(0, END)
+            self.LastName.insert(0, '')
+            self.Score.delete(0, END)
+            self.Score.insert(0, '')
 
-        # autofocusing the first input field
-        self.MatricNo.focus()
+            # autofocusing the first input field
+            self.MatricNo.focus()
 
-        # print(self.matrix)
+            # print(self.matrix)
+
+        else:
+            tkinter.messagebox.showerror("Error", "Student info not complete !")
 
     def clearEntry(self):
         self.MatricNo.delete(0, END)
@@ -271,7 +285,7 @@ class MainPage:
         # print(self.averageScore)
 
         # hiding the frame2 and showing the eda results frame
-        self.frame3.pack()
+        self.frame3.pack(fill="both", expand=True)
         self.frame2.pack_forget()
 
         #clearing any content in the namesLabel label
@@ -293,7 +307,7 @@ class MainPage:
         # print(self.maxScore)
 
         # hiding the frame2 and showing the eda results frame
-        self.frame3.pack()
+        self.frame3.pack(fill="both", expand=True)
         self.frame2.pack_forget()
 
         #clearing any content in the namesLabel label
@@ -315,7 +329,7 @@ class MainPage:
         # print(self.minScore)
 
         # hiding the frame2 and showing the eda results frame
-        self.frame3.pack()
+        self.frame3.pack(fill="both", expand=True)
         self.frame2.pack_forget()
 
         #clearing any content in the namesLabel label
@@ -346,7 +360,7 @@ class MainPage:
         # print(self.highestStudent)
 
         # hiding the frame2 and showing the eda results frame
-        self.frame3.pack()
+        self.frame3.pack(fill="both", expand=True)
         self.frame2.pack_forget()
 
         #empty array to store the text to be displayed for each student with the highest mark
@@ -364,7 +378,7 @@ class MainPage:
         self.namesLabel.destroy()
         
         #creating the label for the students info
-        self.namesLabel = Label(self.frame3, text='')
+        self.namesLabel = Label(self.frame3, text='', font=('Arial 13'), padx=20, pady=10, bg='yellow', fg='black', justify=LEFT)
         self.namesLabel.pack()
         
         #a function to display as many labels as possible for each high score
@@ -393,7 +407,7 @@ class MainPage:
         #print(self.passScores)
 
          # hiding the frame2 and showing the eda results frame
-        self.frame3.pack()
+        self.frame3.pack(fill="both", expand=True)
         self.frame2.pack_forget()
 
         #empty array to stored the text to be displayed for each student above the 70 mark
@@ -438,7 +452,7 @@ class MainPage:
         #print(self.passScores)
 
          # hiding the frame2 and showing the eda results frame
-        self.frame3.pack()
+        self.frame3.pack(fill="both", expand=True)
         self.frame2.pack_forget()
 
         #empty array to stored the text to be displayed for each student below the 40 mark
@@ -468,7 +482,7 @@ class MainPage:
         self.namesLabel['text'] = addFailedNameToLabel()
 
     def backToFrame2Btn(self):
-        self.frame2.pack()
+        self.frame2.pack(fill="both", expand=True)
         self.frame3.pack_forget()
